Split proposal card metadata helpers
